backend/requirements.txt
boto3
flask
python-dotenv
PIL